Use nullNode instead of false for null
Really nullNode should be a singleton but as it stands each node has a
nullNode object for its left and right and bst has one for its use--
initial root, etc.  Because of this (node.asString == "Null node") is
the conditional used rather than (node == nullNode).  So it's a hacky
halfway point to a nullNode singleton.

Also, for some reason nodes are being printed to the console every time
nodes are added or retrieved from bst, despite there being no print
statements around.