Add values() using an iterator
Right now the iterator is hard-coded to return values.  It's a
little unclear at this point if we will need it to also return
keys and/or bindings.  The docs and tests have that functionality
but the Prof's last post says we just need values.  Adding that
shouldn't be too difficult.